+++ src/components/visualisation/jobParams.test.tsx
@@ -2,7 +2,7 @@ import { renderWithProviders } from "utils/test-utils";
 import { screen } from "@testing-library/react";
 import { JobParamsDrawer } from "./jobParams";
 import { server } from "mocks/server";
-import { rest } from "msw";
+import { http, HttpResponse } from "msw";
 
 describe("Job Parameters Drawer", () => {
   it("should display job parameters", async () => {
@@ -14,8 +14,10 @@ describe("Job Parameters Drawer", () => {
 
   it("should display empty list if backend returns 404", async () => {
     server.use(
-      rest.get("http://localhost/processingJob/1/parameters", (req, res, ctx) =>
-        res.once(ctx.status(404), ctx.delay(0))
+      http.get(
+        "http://localhost/processingJob/1/parameters",
+        () => HttpResponse.json({}, { status: 404 }),
+        { once: true }
       )
     );