+++ src/loaders/session.test.ts
@@ -1,5 +1,5 @@
 import { server } from "mocks/server";
-import { rest } from "msw";
+import { http, HttpResponse } from "msw";
 import { queryClient } from "utils/test-utils";
 import { handleGroupClicked, sessionPageLoader } from "loaders/session";
 
@@ -15,8 +15,10 @@ describe("Session Data", () => {
 
   it("should return null if session request fails", async () => {
     server.use(
-      rest.get("http://localhost/proposals/:propId/sessions/:visitId", (req, res, ctx) =>
-        res(ctx.status(404), ctx.delay(0))
+      http.get(
+        "http://localhost/proposals/:propId/sessions/:visitId",
+        () => HttpResponse.json({}, { status: 404 }),
+        { once: true }
       )
     );
     const data = await sessionPageLoader(queryClient)(request, { propId: "1", visitId: "1" });
@@ -27,7 +29,9 @@ describe("Session Data", () => {
 
   it("should return empty list if data collection groups request fails", async () => {
     server.use(
-      rest.get("http://localhost/dataGroups", (req, res, ctx) => res(ctx.status(404), ctx.delay(0)))
+      http.get("http://localhost/dataGroups", () => HttpResponse.json({}, { status: 404 }), {
+        once: true,
+      })
     );
     const data = await sessionPageLoader(queryClient)(request, { propId: "1", visitId: "1" });